todo-app-backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.2
python-multipart==0.0.20
psycopg2-binary==2.9.9
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-dotenv==1.0.1
asyncpg==0.30.0
psutil==6.0.0
prometheus-client==0.20.0